Extensive gridding work, bugs still present.
* Testing shows differences vs. RST, but is able to produce files. Need
  to investigate where the differences arise.

+++ src/utils/channel.rs
@@ -1,6 +1,6 @@
 use crate::error::ProcdarnError;
 
-pub fn set_stereo_channel(channel_char: char) -> Result<i32, ProcdarnError> {
+pub fn set_stereo_channel(channel_char: char) -> Result<i16, ProcdarnError> {
     match channel_char {
         'a' => Ok(1),
         'b' => Ok(2),
@@ -11,7 +11,7 @@ pub fn set_stereo_channel(channel_char: char) -> Result<i32, ProcdarnError> {
     }
 }
 
-pub fn set_fix_channel(channel_char: char) -> Result<i32, ProcdarnError> {
+pub fn set_fix_channel(channel_char: char) -> Result<i16, ProcdarnError> {
     match channel_char {
         'a' => Ok(1),
         'b' => Ok(2),
